sconv-s16be: declaration/implementation mismatches
Fixes the following compiler errors:

./pulsecore/sconv-s16be.h:41:6: warning: type of 'pa_sconv_s24_32be_from_float32ne' does not match original declaration [-Wlto-type-mismatch]
 void pa_sconv_s24_32be_from_float32ne(unsigned n, const float *a, uint8_t *b);
      ^
pulsecore/sconv-s16le.c:413:6: note: 'pa_sconv_s24_32be_from_float32ne' was previously declared here
 void pa_sconv_s24_32le_from_float32ne(unsigned n, const float *a, uint32_t *b) {
      ^
pulsecore/sconv-s16le.c:413:6: note: code may be misoptimized unless -fno-strict-aliasing is used
./pulsecore/sconv-s16be.h:40:6: warning: type of 'pa_sconv_s24_32be_to_float32ne' does not match original declaration [-Wlto-type-mismatch]
 void pa_sconv_s24_32be_to_float32ne(unsigned n, const uint8_t *a, float *b);
      ^
pulsecore/sconv-s16le.c:388:6: note: 'pa_sconv_s24_32be_to_float32ne' was previously declared here
 void pa_sconv_s24_32le_to_float32ne(unsigned n, const uint32_t *a, float *b) {
      ^
pulsecore/sconv-s16le.c:388:6: note: code may be misoptimized unless -fno-strict-aliasing is used
./pulsecore/sconv-s16be.h:56:6: warning: type of 'pa_sconv_s24_32be_from_s16ne' does not match original declaration [-Wlto-type-mismatch]
 void pa_sconv_s24_32be_from_s16ne(unsigned n, const int16_t *a, uint8_t *b);
      ^
pulsecore/sconv-s16le.c:365:6: note: 'pa_sconv_s24_32be_from_s16ne' was previously declared here
 void pa_sconv_s24_32le_from_s16ne(unsigned n, const int16_t *a, uint32_t *b) {
      ^
pulsecore/sconv-s16le.c:365:6: note: code may be misoptimized unless -fno-strict-aliasing is used
./pulsecore/sconv-s16be.h:55:6: warning: type of 'pa_sconv_s24_32be_to_s16ne' does not match original declaration [-Wlto-type-mismatch]
 void pa_sconv_s24_32be_to_s16ne(unsigned n, const uint8_t *a, int16_t *b);
      ^
pulsecore/sconv-s16le.c:342:6: note: 'pa_sconv_s24_32be_to_s16ne' was previously declared here
 void pa_sconv_s24_32le_to_s16ne(unsigned n, const uint32_t *a, int16_t *b) {
      ^
pulsecore/sconv-s16le.c:342:6: note: code may be misoptimized unless -fno-strict-aliasing is used

+++ src/pulsecore/sconv-s16be.h
@@ -37,8 +37,8 @@ void pa_sconv_s24be_from_float32ne(unsigned n, const float *a, uint8_t *b);
 void pa_sconv_s24be_to_float32re(unsigned n, const uint8_t *a, float *b);
 void pa_sconv_s24be_from_float32re(unsigned n, const float *a, uint8_t *b);
 
-void pa_sconv_s24_32be_to_float32ne(unsigned n, const uint8_t *a, float *b);
-void pa_sconv_s24_32be_from_float32ne(unsigned n, const float *a, uint8_t *b);
+void pa_sconv_s24_32be_to_float32ne(unsigned n, const uint32_t *a, float *b);
+void pa_sconv_s24_32be_from_float32ne(unsigned n, const float *a, uint32_t *b);
 void pa_sconv_s24_32be_to_float32re(unsigned n, const uint8_t *a, float *b);
 void pa_sconv_s24_32be_from_float32re(unsigned n, const float *a, uint8_t *b);
 
@@ -52,8 +52,8 @@ void pa_sconv_s24be_from_s16ne(unsigned n, const int16_t *a, uint8_t *b);
 void pa_sconv_s24be_to_s16re(unsigned n, const uint8_t *a, int16_t *b);
 void pa_sconv_s24be_from_s16re(unsigned n, const int16_t *a, uint8_t *b);
 
-void pa_sconv_s24_32be_to_s16ne(unsigned n, const uint8_t *a, int16_t *b);
-void pa_sconv_s24_32be_from_s16ne(unsigned n, const int16_t *a, uint8_t *b);
+void pa_sconv_s24_32be_to_s16ne(unsigned n, const uint32_t *a, int16_t *b);
+void pa_sconv_s24_32be_from_s16ne(unsigned n, const int16_t *a, uint32_t *b);
 void pa_sconv_s24_32be_to_s16re(unsigned n, const uint8_t *a, int16_t *b);
 void pa_sconv_s24_32be_from_s16re(unsigned n, const int16_t *a, uint8_t *b);
